Handle subpackages and remove package
Included the functionality to install packages that originate from a
larger package (e.g. gvim.) Otherwise, some packages would be
uninstallable since the only difference is compile flags.

Further work needs to be done to prevent the extraneous packages from
being compiled as that increases the number of packages to compile. This
may not be possible due to how makepkg functions.

Additionally, a simple remove functionality was included that uses
pacman to remove the package. During the removal process, the git and
package directory are removed to declutter drive space.