Implement and refactor implements
Implemented several improvements to the code to make it more readable
and maintain in addition to adding extra functionality. The list of
features added or refactored are:

 - Change commands to be similar to pacman
 - Simplified installation and update code
 - Update code now updates all packages
 - Created several helper functions to make code more readable
 - Refactored existing functions to use passed arguments

One of the major changes in the overall design and architecture of the
tool is the transition to a more supplemental tool. This way asb can be
used in conjuction with pacman if the user does not wish compile
everything. This allows the tool to be useful to more users.

There are several TODOs within the file that need to be completed prior
to the script can reach a more stable status.